third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-invoking-attribute.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popup invoking attribute</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<button id=b1 popup=p1>Open Popup 1</button>
<div popup=popup id=p1 anchor=b1><p>This is popup #1</p></div>

<style>
  [popup] { border: 5px solid red; }
</style>

<script>
  function clickOn(element) {
    const actions = new test_driver.Actions();
    return actions.pointerMove(0, 0, {origin: element})
      .pointerDown({button: actions.ButtonType.LEFT})
      .pointerUp({button: actions.ButtonType.LEFT})
      .send();
  }

  const popup1 = document.querySelector('#p1');
  const button1 = document.querySelector('#b1');

  promise_test(async () => {
    assert_false(popup1.matches(':popup-open'));
    await clickOn(button1);
    assert_true(popup1.matches(':popup-open'));
    popup1.hidePopup();
    assert_false(popup1.matches(':popup-open'));
    button1.click();
    assert_true(popup1.matches(':popup-open'));
  }, "Basic test of the 'popup' attribute on a button");
</script>
